Fix a crash with a link
The cmark file parser.h was missing a #include. This resulted in the compiler generating bad code for a reference to the mem struct. This caused men->free() to actually call mem->xcalloc() in the reference_free() function, which caused the crash. This has been fixed in the cmark project source code.